chore(tee): gate mock attestation behind default-off mock-attestation feature
Make the mock TEE attestation path structurally absent from production
binaries instead of runtime-guarded. With the feature off, is_mock_quote /
verify_mock_attestation / generate_mock_attestation / MOCK_QUOTE_HEADER do
not exist, so a MOCK_TDX_QUOTE_V1 blob simply fails real DCAP parsing and is
rejected. No mock-detection code is compiled into a release build.

Gated symbols (calimero-tee-attestation, publish = true):
- MOCK_QUOTE_HEADER, generate_mock_attestation, is_mock_quote,
  create_mock_quote, verify_mock_attestation, and their re-exports.
- The non-Linux generate_attestation splits by feature: mock fallback with
  the feature, AttestationError::QuoteGenerationFailed without it.

Feature graph (none in default features):
- calimero-tee-attestation: mock-attestation = []
- calimero-server -> calimero-tee-attestation/mock-attestation
- calimero-node -> calimero-tee-attestation + calimero-server
- merod -> calimero-node + calimero-server + calimero-tee-attestation

Gated call sites (real path is the sole compiled path when off):
- node tee_attestation_admission: fleet admission only ever calls
  verify_attestation.
- node specialized_node_invite: mock branch gated (required to compile; the
  invite path itself is otherwise untouched).
- server tee/attest + tee/fleet_join: mock mint branch and the mock reject.
- merod kms: both mock verify branches, MOCK_KMS_ATTESTATION_ENV,
  mock_kms_attestation_runtime_opt_in, the mock probe-code arm, mock tests.
- merod tee probe: mock flag/branches and mock unit tests.
- node local_governance_node_e2e (+ cascade_dispatch_e2e, which reuses its
  boot harness) gated behind the feature.

mock_tee field plumbing (NodeConfig, AdminState, server::start) is gated so
no field is dead when the feature is off; merod run --mock-tee and its
deny-guard remain. policy.accept_mock / config accept_mock fields are kept
(governance/KMS policy data, inert without mock code).

Env drops (flags are now CLI-only, killing env inheritance):
- env = "MEROD_MOCK_TEE" removed from merod run and merod tee probe.
- MEROD_ALLOW_MOCK_KMS_ATTESTATION is now compiled out with the feature.

CI:
- release.yml fails the build if cargo tree resolves mock-attestation into
  the production merod build.
- ci-checks.yml gains clippy + test steps with --features mock-attestation so
  the mock coverage does not silently drop.
