+++ crates/merod/Cargo.toml
@@ -57,6 +57,16 @@ x509-cert = "0.2.5"
 [build-dependencies]
 calimero-build-utils.workspace = true
 
+[features]
+# DEV/TEST ONLY. Compiles the mock TEE attestation path across merod (KMS mock
+# verify, `merod tee probe --mock-tee`, mock-tee admin wiring). Default OFF, so
+# release `merod` binaries contain no mock-attestation code. See crate docs.
+mock-attestation = [
+    "calimero-node/mock-attestation",
+    "calimero-server/mock-attestation",
+    "calimero-tee-attestation/mock-attestation",
+]
+
 [dev-dependencies]
 axum.workspace = true
 tempfile.workspace = true
