+++ crates/node/src/lib.rs
@@ -45,6 +45,8 @@ pub use run::{start, NodeConfig, NodeMode, SpecializedNodeConfig};
 pub(crate) use state::{CachedBlob, NodeClients, NodeManagers, NodeState};
 pub use sync::SyncManager;
 
+// Not a mock test: uses the feature-ungated `test_node_harness` boot harness,
+// so it runs in the default `cargo test`.
 #[cfg(test)]
 mod cascade_dispatch_e2e;
 #[cfg(test)]
@@ -53,7 +55,12 @@ mod crash_recovery_test;
 mod delta_store_batch_test;
 #[cfg(test)]
 mod delta_store_head_hashes_test;
-#[cfg(test)]
+#[cfg(all(test, feature = "mock-attestation"))]
 mod local_governance_node_e2e;
+/// Feature-ungated (`#[cfg(test)]`-only) in-process node harness shared by
+/// `local_governance_node_e2e` and `cascade_dispatch_e2e`. Contains no mock
+/// attestation code — see the module docs before adding any.
+#[cfg(test)]
+mod test_node_harness;
 #[cfg(test)]
 mod test_support;
