+++ crates/server/Cargo.toml
@@ -47,6 +47,11 @@ calimero-store = { workspace = true, features = ["serde"] }
 calimero-tee-attestation.workspace = true
 mero-auth.workspace = true
 
+[features]
+# DEV/TEST ONLY. Compiles the mock TEE attestation path in the admin TEE
+# handlers (mock quote mint/accept). Default OFF. See crate docs.
+mock-attestation = ["calimero-tee-attestation/mock-attestation"]
+
 [dev-dependencies]
 color-eyre.workspace = true
 tokio = { workspace = true, features = ["macros", "rt-multi-thread"] }
